# swagger.yaml
openapi: 3.0.0
info:
  title: Company & User Service API
  description: |
    API for managing companies, users, Stripe Connect onboarding, and payment transactions.
    This service relies on an API Gateway for authentication, which should pass user and company information via headers.
  version: 1.1.0
servers:
  - url: http://localhost:3000/api
    description: Local Development Server
...
paths:
...
  /companies:
    get:
      tags:
        - Companies
      summary: Retrieve a list of all companies
      description: Fetches all registered companies. Requires admin role.
      security:
        - gatewayAuth: [admin]
      responses:
        '200':
          description: A list of companies.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Company'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags:
        - Companies
      summary: Create a new company
      description: Registers a new company. Requires admin role.
      security:
        - gatewayAuth: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewCompany'
      responses:
        '201':
          description: Company created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  responses:
    BadRequest:
      description: Invalid request payload or parameters.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: "Invalid input provided."
    Unauthorized:
      description: Authentication information is missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: "Authentication required. Please provide valid headers."
    Forbidden:
      description: The authenticated user does not have permission to perform this action.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: "You do not have permission to access this resource."
...
    Conflict:
      description: The request could not be completed due to a conflict with the current state of the resource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: "A resource with this identifier already exists."
  securitySchemes:
    gatewayAuth:
      type: apiKey
      in: header
      name: X-User-ID # This is a placeholder; multiple headers are used.
      description: |
        Authentication is handled by the API Gateway, which must pass the following headers:
        - `X-User-ID`: The UUID of the authenticated user.
        - `X-Company-ID`: The UUID of the user's company.
        - `X-User-Role`: The role of the user (e.g., `admin`, `user`).
  schemas:
...
    Company:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        address:
          type: string
        phone:
          type: string
        stripeAccountId:
          type: string
    NewCompany:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        address:
          type: string
        phone:
          type: string
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: A human-readable error message.
          example: "Company not found"
        error:
          type: string
          nullable: true
          description: Optional detailed error information (e.g., stack trace or specific database error).
          example: "SequelizeUniqueConstraintError: Validation error"